[CATALYST-1743] - Create Project Commands (#2865)
* create project command with three nested commands (create, link, and list)

* updated error message in build command

* fix linting

* Update packages/catalyst/src/cli/commands/build.ts


* merge link tests into project

* organize tests for project using describe

* fix linting

---------
